--- app/src/main/java/com/foodamental/AlertPage.java
@@ -1,64 +1,0 @@
-package com.foodamental;
-
-import android.app.Activity;
-import android.app.AlarmManager;
-import android.app.PendingIntent;
-import android.content.Intent;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
-import android.widget.TextView;
-import android.widget.TimePicker;
-import android.widget.ToggleButton;
-
-import java.util.Calendar;
-
-public class AlertPage extends Activity {
-
-    AlarmManager alarmManager;
-    private PendingIntent pendingIntent;
-    private TimePicker alarmTimePicker;
-    private static AlertPage inst;
-    private TextView alarmTextView;
-
-    public static AlertPage instance() {
-        return inst;
-    }
-
-    @Override
-    public void onStart() {
-        super.onStart();
-        inst = this;
-    }
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.alert_page);
-        alarmTimePicker = (TimePicker) findViewById(R.id.alarmTimePicker);
-        alarmTextView = (TextView) findViewById(R.id.alarmText);
-        ToggleButton alarmToggle = (ToggleButton) findViewById(R.id.alarmToggle);
-        alarmManager = (AlarmManager) getSystemService(ALARM_SERVICE);
-    }
-
-    public void onToggleClicked(View view) {
-        if (((ToggleButton) view).isChecked()) {
-            Log.d("MyActivity", "Alarm On");
-            Calendar calendar = Calendar.getInstance();
-            //calendar.set(Calendar.DAY_OF_YEAR, alarmTimePicker.getCurrentHour());
-          //  calendar.set(Calendar.HOUR_OF_DAY, alarmTimePicker.getCurrentHour());
-           // calendar.set(Calendar.MINUTE, alarmTimePicker.getCurrentMinute());
-            Intent myIntent = new Intent(AlertPage.this, AlarmReceiver.class);
-            pendingIntent = PendingIntent.getBroadcast(AlertPage.this, 0, myIntent, 0);
-            alarmManager.set(AlarmManager.RTC, calendar.getTimeInMillis(), pendingIntent);
-        } else {
-            alarmManager.cancel(pendingIntent);
-            setAlarmText("");
-            Log.d("MyActivity", "Alarm Off");
-        }
-    }
-
-    public void setAlarmText(String alarmText) {
-        alarmTextView.setText(alarmText);
-    }
-}
