+++ app/src/main/java/com/foodamental/activity/Parametres.java
@@ -1,4 +1,4 @@
-package com.foodamental;
+package com.foodamental.activity;
 
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
@@ -6,6 +6,8 @@ import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
+
+import com.foodamental.R;
 
 public class Parametres extends AppCompatActivity {
