+++ app/src/main/java/com/foodamental/activity/Profile.java
@@ -1,22 +1,19 @@
-package com.foodamental;
+package com.foodamental.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
-import android.view.View;
 import android.widget.TextView;
-import android.widget.Toast;
+
+import com.foodamental.R;
+import com.foodamental.util.StaticUtil;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.InputStreamReader;
 
 public class Profile extends AppCompatActivity {
