+++ app/src/main/java/com/foodamental/dao/DBHelper.java
@@ -1,16 +1,12 @@
-package com.foodamental;
+package com.foodamental.dao;
 
-import android.content.ContentValues;
-import android.content.Context;
+
 import android.database.Cursor;
-import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
-import android.database.sqlite.SQLiteDatabaseCorruptException;
 import android.database.sqlite.SQLiteOpenHelper;
-import java.util.ArrayList;
-import java.util.List;
-import android.database.MatrixCursor;
 import android.util.Log;
+
+import com.foodamental.activity.MyMainPage;
 
 import java.util.ArrayList;
