+++ app/src/main/java/com/foodamental/dao/ProductDB.java
@@ -1,13 +1,10 @@
-package com.foodamental;
+package com.foodamental.dao;
 
 import android.content.ContentValues;
-import android.content.Context;
 import android.database.Cursor;
-import android.database.MatrixCursor;
-import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
-import android.database.sqlite.SQLiteOpenHelper;
-import android.util.Log;
+
+import com.foodamental.model.ProductObject;
 
 import java.util.ArrayList;
 import java.util.List;
