+++ app/src/main/java/com/foodamental/dao/UserDB.java
@@ -1,12 +1,10 @@
-package com.foodamental;
+package com.foodamental.dao;
 
 import android.content.ContentValues;
-import android.content.Context;
 import android.database.Cursor;
-import android.database.MatrixCursor;
-import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
-import android.util.Log;
+
+import com.foodamental.model.FoodUser;
 
 import java.util.ArrayList;
 import java.util.List;
