+++ app/src/main/java/com/foodamental/model/ProductObject.java
@@ -1,4 +1,4 @@
-package com.foodamental;
+package com.foodamental.model;
 
 import java.util.List;
