+++ app/src/main/java/com/foodamental/util/RecipeItem.java
@@ -1,4 +1,4 @@
-package com.foodamental;
+package com.foodamental.util;
 
 import android.graphics.Bitmap;
 import android.media.Image;
