+++ app/src/main/java/com/foodamental/util/RecipesArrayAdapter.java
@@ -1,9 +1,6 @@
-package com.foodamental;
+package com.foodamental.util;
 
 import android.content.Context;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.os.AsyncTask;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -11,9 +8,8 @@ import android.widget.ArrayAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import java.io.IOException;
-import java.net.MalformedURLException;
-import java.net.URL;
+import com.foodamental.R;
+
 import java.util.List;
 
 /**
