+++ app/src/main/java/com/foodamental/util/TweetAdapter.java
@@ -1,4 +1,4 @@
-package com.foodamental;
+package com.foodamental.util;
 
 import android.content.Context;
 import android.graphics.drawable.ColorDrawable;
@@ -8,6 +8,8 @@ import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import com.foodamental.R;
 
 import java.util.List;
