+++ app/src/main/res/layout/activity_my_welcome_page.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.foodamental.MyWelcomePage">
+    tools:context=".activity.MyWelcomePage">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
