+++ app/src/main/res/layout/activity_parametres.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.foodamental.Parametres">
+    tools:context=".activity.Parametres">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
