+++ app/src/main/res/layout/activity_recipe_content.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.foodamental.Profile">
+    tools:context=".activity.Profile">
 
     <TextView
         android:layout_width="match_parent"
