+++ app/src/main/res/layout/alert_page.xml
@@ -7,7 +7,7 @@
         android:paddingRight="@dimen/activity_horizontal_margin"
         android:paddingTop="@dimen/activity_vertical_margin"
         android:paddingBottom="@dimen/activity_vertical_margin"
-        tools:context="com.foodamental.MyWelcomePage">
+        tools:context=".activity.MyWelcomePage">
 
 
     <ToggleButton
