+++ app/src/main/res/layout/app_bar_courses.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.foodamental.Courses">
+    tools:context=".activity.AlertPage$Courses">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
