+++ app/src/main/res/layout/app_bar_recipes.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.foodamental.Recipes">
+    tools:context=".activity.Recipes">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
