+++ app/src/main/res/layout/content_my_welcome_page.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.foodamental.MyWelcomePage"
+    tools:context=".activity.MyWelcomePage"
     tools:showIn="@layout/activity_my_welcome_page">
