+++ app/src/main/res/layout/content_profile.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.foodamental.Profile"
+    tools:context=".activity.Profile"
     tools:showIn="@layout/activity_profile">
 
     <TextView
