+++ app/src/main/res/menu/menu_my_welcome_page.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.foodamental.MyWelcomePage">
+    tools:context=".activity.MyWelcomePage">
     <item
         android:id="@+id/action_settings"
         android:orderInCategory="100"
